requirements.txt
openai==0.28.1
fhir.resources
python-dotenv